% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CombinePercRank}
\alias{CombinePercRank}
\title{CombinePercRank}
\usage{
CombinePercRank(perc.rank.list = perc.rank.list)
}
\arguments{
\item{perc.rank.list}{output of GetCentrality()}
}
\value{
dataframe of celltypes and their centrality values
}
\description{
calcuate percentile rank of each centrality measure. for those not in the network, but in the nodes of HNv3, 0 value is assigned.
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
}
