% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_UserGeneset.R
\name{Connectivity}
\alias{Connectivity}
\title{Connectivity}
\usage{
Connectivity(network = NULL, geneset = NULL, simulate.num = 10000)
}
\arguments{
\item{network}{dataframe edgelist of scHumanNet. Must input only one network}

\item{geneset}{character vector of geneset}

\item{simulate.num}{number to simulate the for the null distribution, default is 10,000 times}
}
\value{
A list. First element the null distribution of connectivity from rejection sampling, second element non-parametric pvalue of your geneset, third element the detected genes used for analysis
}
\description{
Deconvolution of geneset connectivity for each scHumanNet constructed
}
\examples{
data("ICMs")
Connectivity(network.list = sorted.net.list, geneset = icm.genes)
}
