% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_UserGeneset.R
\name{DeconvoluteNet}
\alias{DeconvoluteNet}
\title{DeconvoluteNet}
\usage{
DeconvoluteNet(network.list = NULL, geneset = NULL)
}
\arguments{
\item{network.list}{network list of edgelist, this is the output of SortAddLLS}

\item{geneset}{character vector of geneset it can also be a list of multiple geneset}
}
\value{
dataframe of that stores absolute connectivity, connectivty normalized for number of detected geneset in each scHumanNet, and the celltype of scHumanNet. Note: Normalized connectivity should only be used for multiple genesets comparison, and is the value of connectivty devided by the detected number of geneset within each scHumanNet
}
\description{
Deconvolution of geneset connectivity for each scHumanNet constructed
}
\examples{
data("ICMs")
DeconvoluteNet(network.list = sorted.net.list, geneset = icm.genes)
}
