% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DiffPR}
\alias{DiffPR}
\title{DiffPR}
\usage{
DiffPR(
  rank.df.final = NULL,
  celltypes = NULL,
  condition = NULL,
  control = NULL,
  meta = meta
)
}
\arguments{
\item{rank.df.final}{output of CombinePercRank}

\item{celltypes}{column name that stores annoated celltypes}

\item{condition}{column name taht stores disease vs control}

\item{control}{character string that states which of the two condition in the condition column name that will be the control}

\item{meta}{meta.data df that corresponds to each barcode}
}
\value{
dataframe of celltypes and their diffPR values of each genes in the scHumanNet
}
\description{
get Differnence of normalized centrality values from the output of GetCentrality()
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
diffPR.df <- DiffPR(rank.df.final, celltypes = 'celltypes_merged', condition = 'diagnosis', control = 'Control', meta = meta)
}
