% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesetHnv3.R
\name{GenesetHnv3}
\alias{GenesetHnv3}
\title{GenesetHnv3}
\usage{
GenesetHnv3(geneset = NULL, reference.network = graph.hn3)
}
\arguments{
\item{geneset}{geneset of interest. provided multipe genelist as a named list or as a character vector}

\item{reference.network}{default HumanNetv3}
}
\value{
a dataframe of genelist with connectivity, detected genes in the HumanNetv3 and the normalized connectivity
}
\description{
GenesetHnv3
}
\examples{
data("BC_signatures")
hnv3.connectivity.bcsig <- GenesetHnv3(geneset = bc.sig.list)
}
