% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SortAddLLS}
\alias{SortAddLLS}
\title{SortAddLLS}
\usage{
SortAddLLS(Celltype.specific.networks = NULL, reference.network = graph.hnv3)
}
\arguments{
\item{Celltype.specific.networks}{SCINET output from function run.SCINET.clusters().}

\item{reference.network}{reference network input, default is HumanNetv3.}
}
\value{
list of dataframe edgelist, with gene interaction and weights from SCINET and HumanNetv3.
}
\description{
sort SCINET output edges alphabetically and add LLS weight derived from HumanNetv3 reference interactome
}
\examples{
SortAddLLS(Celltype.specific.networks)
}
