% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TopHub}
\alias{TopHub}
\title{TopHub}
\usage{
TopHub(rank.df.final = NULL, top.n = NULL)
}
\arguments{
\item{perc.rank.list}{output of GetCentrality()}
}
\value{
dataframe of celltypes and their top n centrality genes
}
\description{
Get top n genes in terms of centrality for each scHumanNet
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
top.df <- TopHub(rank.df.final, top.n = 50)
}
