% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CombinePercRank}
\alias{CombinePercRank}
\title{CombinePercRank}
\usage{
CombinePercRank(perc.rank.list = perc.rank.list)
}
\arguments{
\item{perc.rank.list}{Output of GetCentrality()}
}
\value{
Dataframe of celltypes and their centrality values
}
\description{
Calcuate percentile rank of each centrality measure. For those not in the network but in the nodes of HNv3, 0 value is assigned.
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
}
