% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DiffPR}
\alias{DiffPR}
\title{DiffPR}
\usage{
DiffPR(
  rank.df.final = NULL,
  celltypes = NULL,
  condition = NULL,
  control = NULL,
  meta = NULL
)
}
\arguments{
\item{rank.df.final}{Output of CombinePercRank}

\item{celltypes}{String character of column name that stores annoated celltypes}

\item{condition}{String character of column name that stores disease vs control}

\item{control}{Character string that states which of the two condition in the condition column name that will be the control}

\item{meta}{meta.data Dataframe that contatins annotated celltypes and condition columns}
}
\value{
Dataframe of celltypes and their diffPR values of each genes in the scHumanNet
}
\description{
Get differnence of normalized centrality values from the output of GetCentrality()
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
diffPR.df <- DiffPR(rank.df.final, celltypes = 'celltypes_merged', condition = 'diagnosis', control = 'Control', meta = meta)
}
