% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{FindDiffHub}
\alias{FindDiffHub}
\title{FindDiffHub}
\usage{
FindDiffHub(diffPR.df = NULL, p.value = NULL)
}
\arguments{
\item{diffPR.df}{Output form diffPR()}

\item{p.value}{Thresholded pvalue, e.g. 0.05 or 0.01}
}
\value{
Dataframe of celltypes and their centrality values
}
\description{
Calcuate statistical signficance of diffPR values calculated with DiffPR()
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list)
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
diffPR.df <- DiffPR(rank.df.final, celltypes = 'celltypes_merged', condition = 'diagnosis', control = 'Control', meta = meta)
diffPR.df.sig <- DiffPR(diffPR.df = diffPR.df, p.value = 0.05)
}
