% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetCentrality}
\alias{GetCentrality}
\title{GetCentrality}
\usage{
GetCentrality(method = method, net.list = NULL)
}
\arguments{
\item{method}{Centrality measure to calculate, supports degree(strength), betweenness, closeness, and eigenvector}

\item{net.list}{Output of SortAddLLS()}
}
\value{
List of named vector, each value corresponding to node's centrality value
}
\description{
Get centrality values for each nodes of scHumanNet list. Ribosomal genes are excluded when calculating centralities
}
\examples{
strenght.list <- GetCentrality(method = 'degree', net.list = Celltype.specific.networks)
}
