% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDiffHub.R
\name{TopDiffHub}
\alias{TopDiffHub}
\title{Title Get top percent of diffHubs default is 5percent}
\usage{
TopDiffHub(diffPR.df = NULL, top.percent = 0.05)
}
\arguments{
\item{diffPR.df}{output dataframe of diffPR(),}

\item{top.percent}{top percentage of diffPR hub genes to retrieve, default is 0.05}
}
\value{
dataframe of top percentage diffPR genes for each celltype
}
\description{
Title Get top percent of diffHubs default is 5percent
}
\examples{
sorted.net.list <- SortAddLLS(Celltype.specific.networks)
strength.list <- GetCentrality(net.list = sorted.net.list, method = 'degree')
rank.df.final <- CombinePercRank(perc.rank.list = strength.list)
diffPR.df <- DiffPR(rank.df.final, celltypes = 'celltypes_merged', condition = 'diagnosis', control = 'Control', meta = meta)
top.diffPR.df <- TopDiffHub(diffPR.df)
}
